added new trigger items, define iso taus with discr., run svfit in parallell, new xsections

+++ TauTauStudies/interface/ElecTauStreamAnalyzer.h
@@ -93,7 +93,7 @@ class ElecTauStreamAnalyzer : public edm::EDAnalyzer{
   std::vector< ROOT::Math::LorentzVector<ROOT::Math::PxPyPzE4D<double> > >* extraElectrons_; 
 
   
-  float run_,event_,lumi_;
+  unsigned long run_,event_,lumi_;
   float sumEt_;
   float chIsoLeg1v1_,nhIsoLeg1v1_,phIsoLeg1v1_;
   float chIsoPULeg1v1_,nhIsoPULeg1v1_,phIsoPULeg1v1_;
